requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pathlib2>=2.3.0 